30% done

+++ AutoPatcher-csharkpy/Form1.cs
@@ -33,11 +33,10 @@ namespace AutoPatcher_csharkpy
 
         private void Form1_Load(object sender, EventArgs e)
         {
-            sconsole.Init("LOG");
-            sconsole.DEBUGMSG("Hey Just click this console window to copy the text to your clipboard");
-            DialogResult dialogResult = MessageBox.Show("Please Read The README It's IMPORTANT\nDo You Want To Open it Now?\nYou Can Open It With A Button In The Program\nOr Just Open It In The Folder Later","ATTENTION",MessageBoxButtons.YesNo,MessageBoxIcon.Error);
 
-            if(dialogResult == DialogResult.Yes)
+            DialogResult dialogResult = MessageBox.Show("Please Read The README It's IMPORTANT\nDo You Want To Open it Now?\nYou Can Open It With A Button In The Program\nOr Just Open It In The Folder Later","ATTENTION",MessageBoxButtons.YesNo,MessageBoxIcon.Error);
+            sconsole.Init("LOG");
+            if (dialogResult == DialogResult.Yes)
             {
                 Process.Start("README.txt");
                 sconsole.LOG("Opened README");
@@ -50,7 +49,7 @@ namespace AutoPatcher_csharkpy
             GTAPATH.Text = checkversion();
             if (GTAPATH.Text.Length != 0)
             {
-                MessageBox.Show("1");
+                MessageBox.Show("GTA Folder Found sucessfully");
                 sconsole.LOG("GTA Folder found successfully");
             }
             else
@@ -60,7 +59,7 @@ namespace AutoPatcher_csharkpy
 
         }
 
-        private void button1_Click(object sender, EventArgs e)
+        private void BTN_AUTOPATCHER_Click(object sender, EventArgs e)
         {
             DialogResult rs = MessageBox.Show("Make sure the game runs and all first-time setup programs have been completed, before executing this script\n Start Patcher?","Attention",MessageBoxButtons.YesNo,MessageBoxIcon.Information);
             //If Messagebox input = yes run the autopatcher
@@ -73,13 +72,46 @@ namespace AutoPatcher_csharkpy
                     sconsole.LOG("Found Steam Version");
                     Thread.Sleep(200);
                     string s = Directory.GetCurrentDirectory();
+                    if(File.Exists(GTAPATH.Text + @"\GTA5.exe") &&
+                        File.Exists(GTAPATH.Text + @"\PlayGTAV.exe") &&
+                        File.Exists(GTAPATH.Text + @"\GTAVLauncher.exe") &&
+                        File.Exists(GTAPATH.Text + @"\steam_api64.dll") &&
+                        File.Exists(GTAPATH.Text + @"\update\update.rpf") )
+                    {
+                        sconsole.LOG("sry one of these files were not found\nGTA5.exe\nPlayGTAV.exe\nGTAVLauncher.exe\nsteam_api64.dll\nupdate.rpf");
+                        return;
+                    }
+                    sconsole.LOG("Backing Up Files");//Backing Up all important files,
                     File.Move(GTAPATH.Text + @"\GTA5.exe", s + @"\Backup\Steam\GTA5.exe");
+                    File.Move(GTAPATH.Text + @"\PlayGTAV.exe", s + @"\Backup\Steam\PlayGTAV.exe");
+                    File.Move(GTAPATH.Text + @"\GTAVLauncher.exe", s + @"\Backup\Steam\GTAVLauncher.exe");
+                    File.Move(GTAPATH.Text + @"\steam_api64.dll", s + @"\Backup\Steam\steam_api64.dll");
+                    File.Move(GTAPATH.Text + @"\update\update.rpf", s + @"\Backup\Steam\update.rpf");
 
 
                 }
                 else
                 {
                     sconsole.LOG("Found Rockstar Version");
+                    sconsole.LOG("Found Steam Version");
+                    Thread.Sleep(200);
+                    string s = Directory.GetCurrentDirectory();
+                    if (!File.Exists(GTAPATH.Text + @"\GTA5.exe") &&
+                        !File.Exists(GTAPATH.Text + @"\PlayGTAV.exe") &&
+                        !File.Exists(GTAPATH.Text + @"\GTAVLauncher.exe") &&
+                        !File.Exists(GTAPATH.Text + @"\update\update.rp6"))
+                    {
+                        sconsole.LOG("sry one of these files were not found\nGTA5.exe\nPlayGTAV.exe\nGTAVLauncher.exe\nupdate.rpf");
+                        goto test;
+                    }
+                    else
+                    {
+                        sconsole.LOG("Backing Up Files");
+                        File.Move(GTAPATH.Text + @"\GTA5.exe", s + @"\Backup\Steam\GTA5.exe");
+                    }
+
+
+
                 }
 
 
@@ -88,7 +120,9 @@ namespace AutoPatcher_csharkpy
             {
                 sconsole.LOG("Messagebox Got A No Or Got Closed");
             }
-            //sconsole.LOG("Sucessfully clicked a button",Color.Red);
+        //sconsole.LOG("Sucessfully clicked a button",Color.Red);
+        test:
+            MessageBox.Show("Please look at the LOG as we couldnt find a specific file");
 
         }
 
@@ -105,6 +139,7 @@ namespace AutoPatcher_csharkpy
                     sconsole.LOG("Found Steam Version");
                     Thread.Sleep(200);
                     string s = Directory.GetCurrentDirectory();
+                    sconsole.LOG("Restoring Files");
                     File.Move(s + @"\Backup\Steam\GTA5.exe", GTAPATH.Text + @"\GTA5.exe");
 
 
@@ -112,6 +147,10 @@ namespace AutoPatcher_csharkpy
                 else
                 {
                     sconsole.LOG("Found Rockstar Version");
+                    Thread.Sleep(200);
+                    string s = Directory.GetCurrentDirectory();
+                    sconsole.LOG("Restoring Files");
+                    File.Move(s + @"\Backup\Steam\GTA5.exe", GTAPATH.Text + @"\GTA5.exe");
                 }
